k8s/resource: Make handlers optional in Event.Handle()
In many cases the subscriber to resource events does not care to handle
some event types (e.g. sync). Make the handler functions optional so
the subscriber doesn't need to implement a no-op handler.

+++ pkg/k8s/resource/event.go
@@ -10,8 +10,11 @@ import (
 
 // Event emitted from resource. One of SyncEvent, UpdateEvent or DeleteEvent.
 type Event[T k8sRuntime.Object] interface {
-	// Handle the event by invoking the right handler.
-	// On error the event is requeued by key for later processing.
+	// Handle the event by invoking the right handler. The handlers can be
+	// nil in which case the event is just marked as processed successfully.
+	//
+	// On error from the event handlers the event is requeued by key for later retrying.
+	// The retry behaviour can be configured with the WithErrorHandler option.
 	//
 	// If you use Handle(), then Done() should not be called.
 	// If you need to process the events in parallel/asynchronously,
@@ -50,7 +53,11 @@ type SyncEvent[T k8sRuntime.Object] struct {
 var _ Event[*corev1.Node] = &SyncEvent[*corev1.Node]{}
 
 func (ev *SyncEvent[T]) Handle(onSync func() error, onUpdate func(Key, T) error, onDelete func(Key, T) error) {
-	ev.Done(onSync())
+	if onSync != nil {
+		ev.Done(onSync())
+	} else {
+		ev.Done(nil)
+	}
 }
 
 // UpdateEvent is emitted when an object has been added or updated
@@ -63,7 +70,12 @@ type UpdateEvent[T k8sRuntime.Object] struct {
 var _ Event[*corev1.Node] = &UpdateEvent[*corev1.Node]{}
 
 func (ev *UpdateEvent[T]) Handle(onSync func() error, onUpdate func(Key, T) error, onDelete func(Key, T) error) {
-	ev.Done(onUpdate(ev.Key, ev.Object))
+	if onUpdate != nil {
+		ev.Done(onUpdate(ev.Key, ev.Object))
+	} else {
+		ev.Done(nil)
+	}
+
 }
 
 // DeleteEvent is emitted when an object has been deleted
@@ -76,5 +88,9 @@ type DeleteEvent[T k8sRuntime.Object] struct {
 var _ Event[*corev1.Node] = &DeleteEvent[*corev1.Node]{}
 
 func (ev *DeleteEvent[T]) Handle(onSync func() error, onUpdate func(Key, T) error, onDelete func(Key, T) error) {
-	ev.Done(onDelete(ev.Key, ev.Object))
+	if onDelete != nil {
+		ev.Done(onDelete(ev.Key, ev.Object))
+	} else {
+		ev.Done(nil)
+	}
 }
